pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "biollm-cls"
version = "0.1.0"
description = "CLS-inspired LLM agent MVP with wake/sleep consolidation"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "torch>=2.4.0",
  "numpy>=1.26.0",
  "wandb>=0.19.0",
  "pyyaml>=6.0.0",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "pydantic>=2.8.0",
  "transformers>=4.45.0",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/biollm_cls"]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]
